refactor(ui): apply PR review nits — split types/constants, use enum, drop inline style
- Move EnvironmentBadgeProps and EnvironmentBadgeConfig to
  EnvironmentBadge.types.ts; tooltip strings and BADGE_CONFIG to
  EnvironmentBadge.constants.ts, per the style guide.
- Replace hardcoded 'production' strings with Environment.Production in
  the OAuth and handlers specs.
- Replace the inline paddingLeft on the InstanceHeader env-badge slot
  with a styled FlexItem in InstanceHeader.styles.ts.

References: #RI-8181
