+++ redisinsight/ui/src/components/environment-badge/EnvironmentBadge.types.ts
@@ -1,0 +1,15 @@
+import React from 'react'
+import { Environment } from 'apiClient'
+
+import { BadgeVariants } from 'uiSrc/components/base/display/badge/RiBadge'
+
+export interface EnvironmentBadgeProps {
+  environment?: Environment
+  dataTestId?: string
+}
+
+export interface EnvironmentBadgeConfig {
+  label: string
+  variant: BadgeVariants
+  tooltip: React.ReactNode
+}
