+++ redisinsight/ui/src/components/environment-badge/index.ts
@@ -1,2 +1,2 @@
 export { EnvironmentBadge } from './EnvironmentBadge'
-export type { EnvironmentBadgeProps } from './EnvironmentBadge'
+export type { EnvironmentBadgeProps } from './EnvironmentBadge.types'
