+++ redisinsight/ui/src/components/instance-header/InstanceHeader.styles.ts
@@ -1,6 +1,8 @@
 import React from 'react'
 import styled, { css } from 'styled-components'
 
+import { FlexItem } from 'uiSrc/components/base/layout/flex'
+
 interface InstanceHeaderContainerProps
   extends React.HTMLAttributes<HTMLDivElement> {
   $isProductionEnv?: boolean
@@ -12,4 +14,8 @@ const productionTint = css`
 
 export const InstanceHeaderContainer = styled.div<InstanceHeaderContainerProps>`
   ${({ $isProductionEnv }) => $isProductionEnv && productionTint}
+`
+
+export const EnvironmentBadgeSlot = styled(FlexItem)`
+  padding-left: 8px;
 `
