+++ redisinsight/ui/src/pages/home/components/databases-list/methods/handlers.spec.ts
@@ -1,3 +1,5 @@
+import { Environment } from 'apiClient'
+
 import {
   TelemetryEvent,
   sendEventTelemetry,
@@ -73,7 +75,7 @@ const instance = {
   id: '1',
   provider: 'LOCALHOST',
   modules: [],
-  environment: 'production',
+  environment: Environment.Production,
 }
 
 describe('databases-list methods/handlers', () => {
@@ -98,7 +100,7 @@ describe('databases-list methods/handlers', () => {
         eventData: expect.objectContaining({
           databaseId: '1',
           provider: 'LOCALHOST',
-          environment: 'production',
+          environment: Environment.Production,
         }),
       }),
     )
